Added the 'component' scripting API.
The component API allows one script to have 'child' scripts, accessed as components. Whilst it's not quite complete, it functions perfectly well, and Xsera now has a testcase for it. The scriptig header now also offers a function to get at the raw lua_State*, in case the need arrives as it did for the component routines.